Only autoclean before production Webpack builds
