+++ webpack/common.js
@@ -1,15 +1,4 @@
-const CleanWebpackPlugin = require('clean-webpack-plugin')
 const { resolve } = require('path')
-
-const pathsToClean = [
-  './formatters/epub/dist',
-  './formatters/html/dist',
-  './formatters/html/fonts'
-]
-
-const cleanOptions = {
-  root: resolve(__dirname, '..')
-}
 
 module.exports = {
   entry: {
@@ -68,8 +57,5 @@ module.exports = {
         ]
       }
     ]
-  },
-  plugins: [
-    new CleanWebpackPlugin(pathsToClean, cleanOptions)
-  ]
+  }
 }
